requirements.txt
django
djangorestframework
django-filter
celery
redis
django-redis
